Python/requirements.txt
fastapi
uvicorn
Pillow
python-multipart
resvg-py